feat: gitignore file and bikeshare
Add the gitignore file containing all data files for the bikeshare project. Purpose is to not upload large data files. There are three csv files which contain data for chicago, new york city, and washington.

Add the bikeshare containing the script which will cycle through the data files and allow the user to make input. This commit will show that the biekshare python script is on the repo.